+++ NEWS.md
@@ -1,3 +1,7 @@
+# dymiumCore (development version)
+
+- Added `sim()` for compiling and executing a microsimulation pipeline.
+
 # dymiumCore 0.1.7
 
 - `transition()` gains `values` field, it expects a named vector where its values will be used to replace the outcomes of the model and update the variable in `attr`. See the example section to learnsmore.