+++ NEWS.md
@@ -4,7 +4,13 @@
 - Added `targets` as an active field. 
 - Fixed `get_transition` didn't use the preprocessing function when the `model` argument is of `Model` class.
 - `get_transition` now removes the dot prefix of variables in entity data. Those variables are considered as derived variables which will not be checked against when new entity data are geting added.
-- add `dsample` for sampling from a discrete distribution, an extension of `base:sample`.
+- Added `dsample` for sampling from a discrete distribution, an extension of `base:sample`.
+- Added `remove_entity` for removing entities from an [Entity] object. 
+- `Entity$get_ids` can get the ids of removed entities using newly added the `include_removed` argument.
+- `Entity` gains a `primary_id` field.
+- `Entity$add` did not check for potential duplicate ids in the removed data.
+- `World$set_time` returns self. The `$start_iter` method may be removed in the near future as it is not quite intuitive. 
+- Fixed the unexpected behaviour when a non `Target` object is passed into target arg of `Transition`. Prior to this fix, the object that got passed into the target argument was used to create a Target object. Hence, this created an expected behaviour, when the Target's get method was called it returned a scaled down value of the target. This may not be the desire default effect by users. 
 
 # dymiumCore 0.1.7